+++ src/app/auth/sign-in/page.tsx
@@ -1,4 +1,4 @@
-import SignInAuthForm from "@/features/auth/sing-in/sign-in-auth-form";
+import SignInAuthForm from '@/features/auth/sign-in/sign-in-auth-form';
 
 const SignInPage = () => {
   return <SignInAuthForm />;
