+++ src/features/auth/sign-in/sign-in-auth-form.tsx
@@ -7,7 +7,7 @@ import { useRouter } from 'next/navigation';
 import AuthInput from '@/features/auth/auth-input';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { schema, SignInFormData } from '@/features/auth/sing-in/data/schema';
+import { schema, SignInFormData } from '@/features/auth/sign-in/data/schema';
 import Image from 'next/image';
 import { AUTH_MESSAGE } from '@/constants/message-constants';
